database/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.32
GeoAlchemy2==0.14.2
